Refactor/update profile

+++ src/main/java/com/be_mxh/service/impl/UserServiceImpl.java
@@ -5,6 +5,7 @@ import com.be_mxh.dto.user.UpdatePasswordRequest;
 import com.be_mxh.dto.user.UpdateProfileRequest;
 import com.be_mxh.dto.user.UserProfileResponse;
 import com.be_mxh.dto.user.UserResponse;
+import com.be_mxh.entity.Role;
 import com.be_mxh.entity.User;
 import com.be_mxh.entity.UserPrincipal;
 import com.be_mxh.exception.BadRequestException;
@@ -82,20 +83,29 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserProfileResponse updateProfile(UpdateProfileRequest req) {
-        User user = userRepository.findById(req.getId())
-                .orElseThrow(() -> new RuntimeException("User not found"));
-        if (!req.getAvatar().isEmpty()) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new UnauthorizedException("Unauthenticated");
+        }
+
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+
+        assert userDetails != null;
+        String username = userDetails.getUsername();
+
+        // 2️⃣ Load user từ DB
+        User user = userRepository.findByUsername(username);
+
+        if (req.getAvatar() != null && !req.getAvatar().isEmpty()) {
             if (user.getAvatarUrl() != null)
                 imageUploadService.delete(user.getAvatarUrl());
             ImageUploadResult imageUploadResult = imageUploadService.upload(req.getAvatar(), "avatars");
             user.setAvatarUrl(imageUploadResult.getUrl());
         }
-        user.setFirstName(req.getFirstName());
-        user.setLastName(req.getLastName());
+        user.setFullName(req.getFullName());
         user.setAddress(req.getAddress());
-        user.setPhone(req.getPhone());
-        user.setDateOfBirth(req.getDateOfBirth());
-        user.setGender(User.Gender.valueOf(req.getGender()));
+        user.setPhone(req.getPhone() != null && req.getPhone().isBlank() ? null : req.getPhone());
         user.setHobby(req.getHobby());
         userRepository.save(user);
         return mapToUserInfoDto(user);
@@ -105,15 +115,6 @@ public class UserServiceImpl implements UserService {
     public UserProfileResponse findById(Long id) {
         User user = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
         return mapToUserInfoDto(user);
-    }
-
-    @Override
-    public UserDetails loadUserById(Long id) {
-        Optional<User> user = userRepository.findById(id);
-        if (user.isEmpty()) {
-            throw new NullPointerException();
-        }
-        return UserPrincipal.build(user.get());
     }
 
     @Override
@@ -178,8 +179,7 @@ public class UserServiceImpl implements UserService {
                 .id(user.getId())
                 .username(user.getUsername())
                 .email(user.getEmail())
-                .firstName(user.getFirstName())
-                .lastName(user.getLastName())
+                .fullName(user.getFullName())
                 .phone(user.getPhone())
                 .avatarUrl(user.getAvatarUrl())
                 .dateOfBirth(user.getDateOfBirth())
@@ -193,8 +193,7 @@ public class UserServiceImpl implements UserService {
         return UserResponse.builder()
                 .id(user.getId())
                 .username(user.getUsername())
-                .firstName(user.getFirstName())
-                .lastName(user.getLastName())
+                .fullName(user.getFullName())
                 .email(user.getEmail())
                 .phone(user.getPhone())
                 .address(user.getAddress())
@@ -208,7 +207,7 @@ public class UserServiceImpl implements UserService {
                 .roles(
                         user.getRoles()
                                 .stream()
-                                .map(role -> role.getName())
+                                .map(Role::getName)
                                 .collect(Collectors.toSet())
                 )
                 .createdAt(user.getCreatedAt())
